social api

+++ app/Http/Controllers/Front/FrontController.php
@@ -98,15 +98,16 @@ class FrontController extends Controller
     public function checkCookies($userid)
     {
         if (isset($_COOKIE['uuid'])) {
-            $userFind = User::where('system_id', $_COOKIE['uuid'])->first();
-            $cart = Cart::where('user_id', $userFind->id)->first();
-            $cart->user_id = $userid;
-            $cart->save();
-
+            $userFind = User::where('system_id', Auth::user()->system_user)->first();
+            if($userFind){
+                $cart = Cart::where('user_id', $userFind->id)->first();
+                if($cart){
+                    $cart->user_id = $userid;
+                    $cart->save();
+                }
+                $userFind->delete();
+            }
             setcookie("uuid", "", time() - 3600);
-
-            $userFind->delete();
-
             return redirect()->route('user.checkout');
         }
     }
